# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart ERP & BI System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [purchaseOrders]
      summary: List purchase orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, partial, received, cancelled]
      responses:
        "200":
          description: List of purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [purchaseOrders]
      summary: Create purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseOrderRequest"
      responses:
        "201":
          description: Purchase order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchaseOrders]
      summary: Get purchase order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderDetail"
    delete:
      operationId: deletePurchaseOrder
      tags: [purchaseOrders]
      summary: Cancel purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /purchase-orders/{id}/receive:
    post:
      operationId: receivePurchaseOrder
      tags: [purchaseOrders]
      summary: Receive items from purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceivePurchaseOrderRequest"
      responses:
        "200":
          description: Items received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
...
  /sales-targets:
    get:
      operationId: listSalesTargets
      tags: [tasks]
      summary: List sales targets
      responses:
        "200":
          description: Sales targets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesTarget"
    post:
      operationId: createSalesTarget
      tags: [tasks]
      summary: Create sales target
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalesTargetRequest"
      responses:
        "201":
          description: Target created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesTarget"
...
  /reports/top-selling:
    get:
      operationId: getTopSellingReport
      tags: [reports]
      summary: Top selling products report
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
            default: month
      responses:
        "200":
          description: Top selling products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    PurchaseOrder:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        orderNumber:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        status:
          type: string
          enum: [pending, partial, received, cancelled]
        totalAmount:
          type: number
        discountPercent:
          type: number
        taxPercent:
          type: number
        netAmount:
          type: number
        notes:
          type: string
          nullable: true
        createdByUserId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, orderNumber, supplierId, supplierName, status, totalAmount, discountPercent, taxPercent, netAmount, createdAt]
    PurchaseOrderItem:
      type: object
      properties:
        id:
          type: integer
        purchaseOrderId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productCode:
          type: string
        orderedQuantity:
          type: number
        receivedQuantity:
          type: number
        unitPrice:
          type: number
        totalPrice:
          type: number
      required: [id, purchaseOrderId, productId, productName, productCode, orderedQuantity, receivedQuantity, unitPrice, totalPrice]
    PurchaseOrderDetail:
      allOf:
        - $ref: "#/components/schemas/PurchaseOrder"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/PurchaseOrderItem"
          required: [items]
    CreatePurchaseOrderRequest:
      type: object
      properties:
        supplierId:
          type: integer
        taxPercent:
          type: number
        notes:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItemRequest"
      required: [supplierId, items]
    PurchaseOrderItemRequest:
      type: object
      properties:
        productId:
          type: integer
        orderedQuantity:
          type: number
        unitPrice:
          type: number
      required: [productId, orderedQuantity, unitPrice]
    ReceivePurchaseOrderRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReceiveItemRequest"
      required: [items]
    ReceiveItemRequest:
      type: object
      properties:
        itemId:
          type: integer
        receivedQuantity:
          type: number
      required: [itemId, receivedQuantity]
...
    SalesTarget:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productCode:
          type: string
        currentStock:
          type: number
        targetQuantity:
          type: number
        soldQuantity:
          type: number
        progressPercent:
          type: number
        period:
          type: string
          enum: [week, month]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, productId, productName, productCode, currentStock, targetQuantity, soldQuantity, progressPercent, period, startDate, endDate, createdAt]
    CreateSalesTargetRequest:
      type: object
      properties:
        productId:
          type: integer
        targetQuantity:
          type: number
        period:
          type: string
          enum: [week, month]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required: [productId, targetQuantity, period, startDate, endDate]
...
    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        productCode:
          type: string
        totalQuantity:
          type: number
        totalRevenue:
          type: number
        orderCount:
          type: integer
      required: [productId, productName, productCode, totalQuantity, totalRevenue, orderCount]
...